Telerik controls should use Metro skin
The Simple skin was removed from newer versions of Telerik (see http://blogs.telerik.com/aspnet-ajax/posts/13-04-11/6-telerik-asp.net-ajax-skins-going-obsolete)
The Metro skin is based on the Simple skin (i.e. without lots of extra images,
simpler to override), so it should be used as the replacement.  The Default
skin that DNN uses has some broken icons (specifically in the grid export
buttons), so it's best if a skin is specified.